 \name{bladderEset} \docType{data} \alias{bladderEset} \title{Bladder Gene Expression Data Illustrating Batch Effects} \description{ This is an expression set consisting of 57 samples drawn from a study of bladder cancer. The samples were collected on different dates which have been used to define 5 batches. These data are used asan example in the sva package vignette.  }\references{	Dyrskjot L., Kruhoffer, M. Thykjaer, T., Marcussen N., Jensen J.L., Moller, K., Orntoft, T.F. 2004 \emph{Cancer Research} \bold{64}, 4040--4048.	Leek,J.T., Scharpf, R.B., Corrada Bravo, H., Simcha, D., Langmead, B., Johnson, W.E., Geman, D., Baggerly, K., Irizarry, R.A. 2010 \emph{Nature Reviews Genetics} \bold{11} 733--739} \usage{bladderEset} \keyword{data}